CalReg's save() returns id of saved bean.
And the text ui makes use of this change to fix a bug in adding a new
event to the calendar. editEvent pushes the user back to viewEvent, but
if the id doesn't exist in cache or database, an exception is caught.
Now editEvent() is (should be) guaranteed to send an existing id.